refactor: SearchBar 리팩토링
