+++ frontend/src/components/Header/SearchBar/SelectItemArea/SelectItem/index.ts
@@ -1,0 +1,5 @@
+import CheckInOut from "./CheckInOut";
+import ReservationFee from "./ReservationFee";
+import SelectItem from "./SelectItem";
+
+export { CheckInOut, SelectItem, ReservationFee };
